requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
shap
scikit-learn
xgboost
Pillow
python-magic
python-dotenv
asyncpg
pytest
pytest-asyncio
httpx
